Criando entidades, repositórios e ajeitando banco de dados

+++ src/main/kotlin/com/guiodes/pulsecare/repository/BaseRepository.kt
@@ -7,5 +7,5 @@ interface BaseRepository<T> {
     fun saveOrUpdate(t: T): T
     fun findById(id: Long): T
     fun findAll(): List<T>
-    fun delete(id: Int)
+    fun delete(id: Long)
 }
